Move json_view helper into `h.util.view`
