+++ h/util/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from h.util import user
+from h.util import view
 
-__all__ = ('user',)
+__all__ = ('user', 'view')
